% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_date_first_clean}
\alias{vector_parse_date_first_clean}
\title{cleans incoming dates a little bit so they are a little similar and we don't need to try as many formats}
\usage{
vector_parse_date_first_clean(
  dts,
  TIME_SPLIT = "T",
  seps = "[-.:/\\\\s+]",
  replace_sep = "-"
)
}
\arguments{
\item{dts}{vecotor of date strings}

\item{TIME_SPLIT}{where to split to get rid of the time}

\item{seps}{what the date separators might be}

\item{replace_sep}{what to replace the seperators with}
}
\description{
cleans incoming dates a little bit so they are a little similar and we don't need to try as many formats
}
\examples{
 vector_parse_date_first_clean( c("03/03/92", "03/21/94", "03/02/99", "03/07/02"))
 vector_parse_date_first_clean(dts = c("2018-11-01 08:30:00", "2017-09-19 08:30:00", "2017-02-28 08:30:00"), TIME_SPLIT = " ")
 vector_parse_date_first_clean(dts = c("2018-11-01T08:30:00", "2017-09-19T08:30:00", "2017-02-28T08:30:00"), TIME_SPLIT = "T")
}
