% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_date_is_two_digit_year}
\alias{vector_parse_date_is_two_digit_year}
\title{return true if it is 4 digit year}
\usage{
vector_parse_date_is_two_digit_year(dt_str, fmt, sep = "-")
}
\description{
return true if it is 4 digit year
}
