% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_date_last_120_yr}
\alias{vector_parse_date_last_120_yr}
\title{Can be passed to multiple functions as the 'check_func' parameter,
true if the date is in the past 120 years}
\usage{
vector_parse_date_last_120_yr(dt)
}
\arguments{
\item{dt}{a POSIXct object}
}
\description{
Can be passed to multiple functions as the 'check_func' parameter,
true if the date is in the past 120 years
}
\examples{
  vector_parse_date_last_120_yr(as.Date("2021-01-17"))
  vector_parse_date_last_120_yr(as.Date("2037-01-17"))
  vector_parse_date_last_120_yr(as.Date("0473-7-12"))
}
