% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_date_not_future}
\alias{vector_parse_date_not_future}
\title{true if the date is in the past}
\usage{
vector_parse_date_not_future(dt)
}
\arguments{
\item{dt}{a POSIXct object}
}
\description{
true if the date is in the past
}
