% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_date_only_one}
\alias{vector_parse_date_only_one}
\title{Returns either NA or a date depending on if the dt_str and fmt make sense as a possible date, along with the business logic of check_func}
\usage{
vector_parse_date_only_one(
  dt_str,
  fmt,
  check_func = vector_parse_date_not_future,
  ...
)
}
\arguments{
\item{dt_str}{a single string that may be a date}

\item{fmt}{a single format to try}

\item{check_func}{function return 1 if the date parsed matches business logic, or a decimal for probability.}

\item{...}{passed to check_func

#' @example
   vector_parse_date_only_one(dt_str="03/03/92", fmt="mdy")
   vector_parse_date_only_one(dt_str="03/03/92", fmt="ymd")
   vector_parse_date_only_one(dt_str="20110401", fmt = "dmy", check_func =vector_parse_date_last_120_yr )}
}
\description{
Returns either NA or a date depending on if the dt_str and fmt make sense as a possible date, along with the business logic of check_func
}
