% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_date_usually_true}
\alias{vector_parse_date_usually_true}
\title{Can be passed to multiple functions as the 'check_func' parameter,
always returns true}
\usage{
vector_parse_date_usually_true(dt)
}
\arguments{
\item{dt}{a POSIXct object}
}
\description{
Can be passed to multiple functions as the 'check_func' parameter,
always returns true
}
\examples{
  vector_parse_date_usually_true(as.Date("2021-01-17"))
  vector_parse_date_usually_true(as.Date("2037-01-17"))
  vector_parse_date_usually_true(as.Date("1282-03-27"))
}
