% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_dates}
\alias{vector_parse_dates}
\title{given a vector of dates this will try to parse them using vector of formats, and then parse them as best it can using information on if the other dates parse in a given format}
\usage{
vector_parse_dates(
  dts,
  fmts = vector_parse_date_formats(),
  check_func = vector_parse_date_not_future,
  ...
)
}
\arguments{
\item{dts}{vector of strings}

\item{fmts}{optional vector of formats to check}

\item{check_func}{function that takes a date and returns true if if matches some business logic for a valid date}
}
\description{
given a vector of dates this will try to parse them using vector of formats, and then parse them as best it can using information on if the other dates parse in a given format
}
