% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_dates_guess_at_format}
\alias{vector_parse_dates_guess_at_format}
\title{returns a string for the best guess date, if there is no best guess it return an empty string}
\usage{
vector_parse_dates_guess_at_format(
  dts,
  fmts = vector_parse_date_formats(),
  check_func = vector_parse_date_not_future,
  ...
)
}
\arguments{
\item{dts}{vector of strings}

\item{fmts}{optional vector of formats to check}

\item{check_func}{function that takes a date and returns 1 if if matches some business logic for a valid date}
}
\description{
returns a string for the best guess date, if there is no best guess it return an empty string
}
