% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_dates_guess_at_format_lazy}
\alias{vector_parse_dates_guess_at_format_lazy}
\title{returns a tibble with formats tried and success rate of format in aggregate}
\usage{
vector_parse_dates_guess_at_format_lazy(
  dts,
  fmts = vector_parse_date_formats(),
  check_func = vector_parse_date_not_future,
  origin = "1970-01-01",
  tz = "GMT",
  ...
)
}
\arguments{
\item{dts}{vector of strings}

\item{fmts}{optional vector of formats to check}

\item{check_func}{optional function that takes a date and returns 1 if if matches some business logic for a valid date}

\item{origin}{Origin for postix conversion}

\item{tz}{timezone for postix conversion}

\item{...}{passed to vector_parse_date_only_one ,the nto check_func}
}
\description{
returns a tibble with formats tried and success rate of format in aggregate
}
\examples{
 vector_parse_dates_guess_at_format_lazy(dts=c("03/03/92", "03/21/94", "03/02/99", "03/07/02"))


}
