% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_recent_better}
\alias{vector_parse_recent_better}
\title{Can be passed to multiple functions as the 'check_func' parameter,
Returns a larger number for dates closer to the current date}
\usage{
vector_parse_recent_better(dt)
}
\arguments{
\item{dt}{a POSIXct object}
}
\description{
Can be passed to multiple functions as the 'check_func' parameter,
Returns a larger number for dates closer to the current date
}
\examples{
  vector_parse_recent_better(as.Date("2021-01-17"))

}
