% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_success_rates}
\alias{vector_parse_success_rates}
\title{returns a tibble with formats tried and success rate of format in aggregate}
\usage{
vector_parse_success_rates(
  dts,
  fmts = vector_parse_date_formats(),
  check_func = vector_parse_date_not_future,
  ...
)
}
\arguments{
\item{dts}{vector of strings}

\item{fmts}{optional vector of formats to check}

\item{check_func}{function that takes a date and returns 1 if if matches some business logic for a valid date}
}
\description{
returns a tibble with formats tried and success rate of format in aggregate
}
\examples{
  vector_parse_success_rates(dts=c("03/03/92", "03/21/94", "03/02/99", "03/07/02"))
}
