% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_date_only_one}
\alias{vector_parse_date_only_one}
\title{returns either NA or a date depending on if the dt_str and fmt make sense as a possible date}
\usage{
vector_parse_date_only_one(
  dt_str,
  fmt,
  check_func = vector_parse_date_not_future,
  ...
)
}
\arguments{
\item{dt_str}{a single string that may be a date}

\item{fmt}{a single format to try}
}
\description{
returns either NA or a date depending on if the dt_str and fmt make sense as a possible date
}
