% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_date_possibl_correct_format}
\alias{vector_parse_date_possibl_correct_format}
\title{Return true if the string makes sense as a date in that format}
\usage{
vector_parse_date_possibl_correct_format(
  dt_str,
  fmt,
  check_func = vector_parse_date_not_future,
  ...
)
}
\arguments{
\item{dt_str}{a single string that may be a date}

\item{fmt}{a single format to try}
}
\description{
Return true if the string makes sense as a date in that format
}
