% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_parse_date.R
\name{vector_parse_date_usually_true}
\alias{vector_parse_date_usually_true}
\title{always returns true}
\usage{
vector_parse_date_usually_true(dt)
}
\arguments{
\item{dt}{a POSIXct object}
}
\description{
always returns true
}
